requirements.txt
paho-mqtt
APScheduler==3.10.0
service-bootstrap~=0.0.5